fix(标准产监/决策辅助):定时任务传参报错修改

+++ business/business-decision-support/src/main/java/com/yjtech/wisdom/tourism/decisionsupport/business/vo/AnalyzeDecisionWarnVo.java
@@ -18,5 +18,5 @@ public class AnalyzeDecisionWarnVo implements Serializable {
     /**
      * is_simulation 是否有模拟数据 1有 0无
      */
-    private Byte isSimulation = 0;
+    private Integer isSimulation = 0;
 }
